switch to using twemoji as the source data
Rather than going through emoji-data and looking for the corresponding
twemoji file, go the other way around. There seem to be a bunch of
mistakes in the emoji-data, or at least it's hard to find the
corresponding twemoji filenames. Going this direction yields better
matches between the short names and the images.